05_qrcode_app/requirements.txt
asgiref==3.9.1
Django==5.2.4
mysqlclient==2.2.7
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
pyzbar==0.1.9
qrcode==8.2
sqlparse==0.5.3
